Require logger for Rails 7.0
https://github.com/rails/rails/issues/54263
https://github.com/rails/rails/issues/54260

concurrent-ruby 1.3.5 changes this for Rails, because concurrent-ruby
used to load the logger gem. 1.3.5 no longer does, and Rails needs it.

This should be fixed by newer versions of Rails (include on the 7.0
series), but in the meantime, this fixes the issue. We should remove
this require once we upgrade Rails.

https://github.com/rails/rails/pull/54264